auth/requirements.txt
flask==2.0.2
pymongo==3.12.1
prometheus-flask-exporter==0.18.5
pyyaml==6.0
flask-jwt-extended==4.3.1
redis==3.5.3